+++ lib/blizzard_api/diablo/game_data/generic_data_endpoint.rb
@@ -4,8 +4,8 @@ module BlizzardApi
   module Diablo
     # Generic endpoint to support most data requests with minor configurations
     class GenericDataEndpoint < Diablo::Request
-      def initialize(region = nil)
-        super region
+      def initialize(region = nil, mode = :regular)
+        super region, mode
         endpoint_setup
         @ttl ||= CACHE_DAY
       end