+++ lib/blizzard_api/hearthstone/request.rb
@@ -6,8 +6,8 @@ module BlizzardApi
     class Request < BlizzardApi::Request
       ##
       # @!macro regions
-      def initialize(region = nil)
-        super region
+      def initialize(region = nil, mode = :regular)
+        super region, mode
         @game = 'hearthstone'
       end
     end