+++ test/blizzard/wow/game_data/auction_test.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'test_helper'
+require 'date'
 
 module BlizzardApi
   module Wow
@@ -12,6 +13,16 @@ module BlizzardApi
       def test_auction_get
         auction_data = @auction.get 1146
         assert auction_data.key? :auctions
+      end
+
+      def test_auction_modified_get
+        auction_data = @auction.get 1146, since: DateTime.parse('2000-01-01Z')
+        assert auction_data.key? :auctions
+      end
+
+      def test_auction_unmodified_get
+        auction_data = @auction.get 1146, since: DateTime.parse('2099-01-01Z')
+        assert_nil auction_data
       end
     end
   end